Move `check_receipts_in_primary_block()` to DomainBlockProcessor
